% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_central.R
\name{get_central}
\alias{get_central}
\title{Compute central tendency statistics}
\usage{
get_central(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[base]{mean}} and
\code{\link[stats]{median}}.}
}
\value{
A data frame with two columns:
\describe{
\item{Statistic}{Statistic name.}
\item{Value}{Calculated value for the statistic.}
}
}
\description{
Calculates the mean, median, and mode for a numeric vector.
}
\examples{
get_central(c(1, 2, 2, 3, 4, NA))
}
