% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv.R
\name{get_cv}
\alias{get_cv}
\title{Compute coefficient of variation}
\usage{
get_cv(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[base]{mean}} and
\code{\link[stats]{sd}}.}
}
\value{
The coefficient of variation (percentage).
}
\description{
Calculates the coefficient of variation (CV) of a numeric vector
as a percentage: \code{100 * sd(x) / mean(x)}.
}
\examples{
get_cv(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
