% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iqr.R
\name{get_iqr}
\alias{get_iqr}
\title{Compute IQR and outlier fences}
\usage{
get_iqr(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[stats]{quantile}} and
\code{\link[stats]{IQR}}.}
}
\value{
A data frame with two columns:
\describe{
\item{Statistic}{Statistic name.}
\item{Value}{Calculated value for the statistic.}
}
}
\description{
Calculates quartiles, median, IQR, and Tukey's inner and outer fences.
}
\examples{
get_iqr(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
