% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kurtosis.R
\name{get_kurtosis}
\alias{get_kurtosis}
\title{Compute kurtosis}
\usage{
get_kurtosis(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[base]{mean}} and
\code{\link[stats]{sd}}.}
}
\value{
The kurtosis coefficient.
}
\description{
Calculates the kurtosis of a numeric vector using the sample
formula (fourth standardized moment).
}
\examples{
get_kurtosis(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
