% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs.R
\name{get_obs}
\alias{get_obs}
\title{Compute observation and missing value statistics}
\usage{
get_obs(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A data frame with two columns:
\describe{
\item{Metric}{Statistic name.}
\item{Value}{Calculated value for the statistic.}
}
}
\description{
Calculates the total number of observations, missing values,
non-missing values, and their corresponding percentages for a vector.
}
\examples{
get_obs(c(1, 2, 3, NA))
}
