% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Compute range, minimum and maximum statistics}
\usage{
get_range(x, ...)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A data frame with two columns:
\describe{
\item{Statistic}{Statistic name.}
\item{Value}{Calculated value for the statistic.}
}
}
\description{
Calculates the minimum, maximum, and range for a numeric vector.
}
\examples{
get_range(c(1, 5, 3, 9, 2))
}
