% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sd.R
\name{get_sd}
\alias{get_sd}
\title{Compute standard deviation}
\usage{
get_sd(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[stats]{sd}}.}
}
\value{
A data frame with two columns:
\describe{
\item{Statistic}{Statistic name.}
\item{Value}{Calculated value for the statistic.}
}
}
\description{
Calculates the standard deviation of a numeric vector.
}
\examples{
get_sd(c(1, 2, 3, 4, 5))
}
