% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skewness.R
\name{get_skewness}
\alias{get_skewness}
\title{Compute skewness}
\usage{
get_skewness(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[base]{mean}} and
\code{\link[stats]{sd}}.}
}
\value{
The skewness coefficient.
}
\description{
Calculates the skewness of a numeric vector using the sample
formula (third standardized moment).
}
\examples{
get_skewness(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
