% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\title{Compute variance}
\usage{
get_var(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[stats]{var}}.}
}
\value{
A data frame with two columns:
\describe{
\item{Statistic}{Statistic name.}
\item{Value}{Calculated value for the statistic.}
}
}
\description{
Calculates the variance of a numeric vector.
}
\examples{
get_var(c(1, 2, 3, 4, 5))
}
