% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_sd.R
\name{get_var_sd}
\alias{get_var_sd}
\title{Compute variance and standard deviation}
\usage{
get_var_sd(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[stats]{var}} and
\code{\link[stats]{sd}}.}
}
\value{
A data frame with two columns:
\describe{
\item{Statistic}{Statistic name.}
\item{Value}{Calculated value for the statistic.}
}
}
\description{
Calculates both variance and standard deviation of a numeric vector.
}
\examples{
get_var_sd(c(1, 2, 3, 4, 5))
}
