% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zscore.R
\name{get_zscore}
\alias{get_zscore}
\title{Compute z-scores}
\usage{
get_zscore(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments passed to \code{\link[base]{scale}}.}
}
\value{
The z-score matrix.
}
\description{
Calculates the z-scores (standardized values) of a numeric vector
using \code{\link[base]{scale}}.
}
\examples{
get_zscore(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
