% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{include}
\alias{include}
\title{include function}
\usage{
include(package)
}
\arguments{
\item{package}{library name}
}
\value{
library install and load
}
\description{
include function
}
\examples{
include("ggplot2")
include("dplyr")
}
