% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Print session info to a file}
\usage{
info(file = "info.txt")
}
\arguments{
\item{file}{A character string specifying the name of the file to save
session info. Default is "info.txt".}
}
\value{
A file with session info
}
\description{
Print session info to a file
}
\examples{
info()
}
