% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package.R
\name{is_package}
\alias{is_package}
\title{Check Whether an R Package Is Installed}
\usage{
is_package(x)
}
\arguments{
\item{x}{A character string containing the name of the package.}
}
\value{
A logical value:
\itemize{
\item \code{TRUE} if the package is installed.
\item \code{FALSE} otherwise.
}
}
\description{
Returns \code{TRUE} if the specified package is installed and available,
and \code{FALSE} otherwise.
}
\examples{
is_package("dplyr")
is_package("haven")
}
