% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_clean.R
\name{occ_clean}
\alias{occ_clean}
\title{Clean Occurrence Records from GBIF}
\usage{
occ_clean(
  data,
  coord_uncertainty = 1000,
  occ_type = c("PhysicalObject", "specimen", "PRESERVED_SPECIMEN", "PreservedSpecimen",
    "Occurrence", "Collection", "Physical Object", "Objeto fÃsico"),
  crs_chosen = "albersNA",
  occ_issue = c("COORDINATE_ROUNDED", "RECORDED_DATE_INVALID",
    "COORDINATE_UNCERTAINTY_METERS_INVALID", "PRESUMED_NEGATED_LONGITUDE")
)
}
\arguments{
\item{data}{The file path to species occurrence records downloaded from GBIF}

\item{coord_uncertainty}{The maximum uncertainty around the location of an occurrence.}

\item{occ_type}{The types of occurrence records to be included.}

\item{crs_chosen}{The chosen projection for the occurrences. This parameter is useful if mapping occurrences.}

\item{occ_issue}{A list of species occurrence issues that should not be included in the final data.}
}
\value{
An sf object with clean GBIF data.
}
\description{
Clean Occurrence Records from GBIF
}
