% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup function to install and load libraries}
\usage{
setup(
  packages = c("lobstr", "tidyverse", "haven", "stringr", "DBI", "odbc", "lubridate",
    "ggplot2", "ggalluvial", "RMySQL")
)
}
\arguments{
\item{packages}{a vector of library names to install and load}
}
\value{
libraries installed and loaded
}
\description{
Setup function to install and load libraries
}
\examples{
setup()
setup(c("ggplot2", "dplyr"))
setup("haven")
}
