% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_connect_db.R
\name{sql_connect_db}
\alias{sql_connect_db}
\title{Connect to MySQL database}
\usage{
sql_connect_db(passw, ...)
}
\arguments{
\item{passw}{password for MySQL database connection}

\item{...}{additional parameters for MySQL database connection,
such as dbname, host, port, user, etc.}
}
\value{
a connection object to the MySQL database
}
\description{
Connect to MySQL database
}
\examples{
sql_connect_db(passwd())
}
