% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_load_db.R
\name{sql_load_db}
\alias{sql_load_db}
\title{Load data into MySQL database}
\usage{
sql_load_db(path, ...)
}
\arguments{
\item{path}{the path to the directory containing the data files}

\item{...}{additional parameters for loading data into the MySQL database,
such as file extension, database name, etc.}
}
\value{
no return value, the function writes data to the MySQL database
}
\description{
Load data into MySQL database
}
\examples{
sql_load_db(path = "path/to/data", ext = "xpt", dbname = "mysql")
}
