% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{test}
\alias{test}
\title{Test function plot cars dataset}
\usage{
test(x = cars)
}
\arguments{
\item{x}{dataset name}
}
\value{
A plot cars dataset
}
\description{
Test function plot cars dataset
}
\examples{
test()
}
