% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_tmp.R
\name{test_tmp}
\alias{test_tmp}
\title{Test function plot cars dataset}
\usage{
test_tmp(x = cars)
}
\arguments{
\item{x}{dataset name}
}
\value{
A plot cars dataset
}
\description{
Test function plot cars dataset
}
\examples{
test_tmp()
}
