% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try2leer.R
\name{try2leer}
\alias{try2leer}
\title{try2leer - A generic function to read various file formats and database tables}
\usage{
try2leer(x, ...)
}
\arguments{
\item{x}{the path to the file to read, can be a local file path}

\item{...}{additional arguments passed to the specific read method (e.g., read_csv, read_xpt, etc.)}
}
\value{
a data frame or tibble containing the contents of the file
}
\description{
try2leer - A generic function to read various file formats and database tables
}
\examples{
try2leer("data.csv")
try2leer("data.tsv")
try2leer("data.csv2")
try2leer("data.csv.gz")
try2leer("data.xpt")
try2leer("test.sql")
try2leer("test.db")
}
