% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_input.R
\name{ask_input}
\alias{ask_input}
\title{Ask for user input if variable is NULL}
\usage{
ask_input(var, pprompt, default = NULL)
}
\arguments{
\item{var}{variable to store user input}

\item{pprompt}{prompt message to ask user input}

\item{default}{default value if user input is NULL}
}
\value{
user input or default value
}
\description{
Ask for user input if variable is NULL
}
\examples{
ask_input(a, "Please enter a value: ", default = NULL)
}
