% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCargar.R
\name{DataCargar}
\alias{DataCargar}
\title{Data loader class}
\description{
Load data from files, directories, archives, or databases.
}
\examples{
# SQL example
# db <- DataCargar$new("data.db", db = "sqlite")
# db$leer("my_table")
# db$disconnect()
# zip example
# db <- DataCargar$new("data.zip", db = "zip")
# db$leer("file_in_zip.csv")
# db$disconnect()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datos}}{type of data source (character: "sqlite", "mysql", "zip", etc.)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataCargar-new}{\code{DataCargar$new()}}
\item \href{#method-DataCargar-leer}{\code{DataCargar$leer()}}
\item \href{#method-DataCargar-write}{\code{DataCargar$write()}}
\item \href{#method-DataCargar-guess_ext}{\code{DataCargar$guess_ext()}}
\item \href{#method-DataCargar-disconnect}{\code{DataCargar$disconnect()}}
\item \href{#method-DataCargar-list}{\code{DataCargar$list()}}
\item \href{#method-DataCargar-info}{\code{DataCargar$info()}}
\item \href{#method-DataCargar-clone}{\code{DataCargar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCargar-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataCargar-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DataCargar object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCargar$new(x, db = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file, directory, archive, or a DBIConnection object}

\item{\code{db}}{database}

\item{\code{...}}{Additional arguments passed to internal methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCargar-leer"></a>}}
\if{latex}{\out{\hypertarget{method-DataCargar-leer}{}}}
\subsection{Method \code{leer()}}{
Read data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCargar$leer(table_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{for databases or filename}

\item{\code{...}}{Additional arguments passed to read functions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble/data.frame/lazy table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCargar-write"></a>}}
\if{latex}{\out{\hypertarget{method-DataCargar-write}{}}}
\subsection{Method \code{write()}}{
Write data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCargar$write(df, table_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe or tibble}

\item{\code{table_name}}{name}

\item{\code{...}}{Additional arguments passed to read functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCargar-guess_ext"></a>}}
\if{latex}{\out{\hypertarget{method-DataCargar-guess_ext}{}}}
\subsection{Method \code{guess_ext()}}{
Guess file extension
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCargar$guess_ext(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{path or file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCargar-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-DataCargar-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect from DB if needed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCargar$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCargar-list"></a>}}
\if{latex}{\out{\hypertarget{method-DataCargar-list}{}}}
\subsection{Method \code{list()}}{
List tables or files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCargar$list(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to read functions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCargar-info"></a>}}
\if{latex}{\out{\hypertarget{method-DataCargar-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCargar$info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The DataCargar object itself (invisibly), allowing for method chaining # nolint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCargar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataCargar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCargar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
