% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStat.R
\name{GetStat}
\alias{GetStat}
\title{Statistical analysis class}
\description{
Compute descriptive statistics, observation counts,
missing values, and frequency tables for a vector.
}
\examples{
# Numeric vector
s <- GetStat$new(c(1, 2, 3, NA, 5, 5, 6))
s$get_obs()


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x}}{Input vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetStat-new}{\code{GetStat$new()}}
\item \href{#method-GetStat-obs}{\code{GetStat$obs()}}
\item \href{#method-GetStat-list}{\code{GetStat$list()}}
\item \href{#method-GetStat-clone}{\code{GetStat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStat-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetStat-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GetStat object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStat$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStat-obs"></a>}}
\if{latex}{\out{\hypertarget{method-GetStat-obs}{}}}
\subsection{Method \code{obs()}}{
Observation and missing value statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStat$obs(col = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{column names}

\item{\code{...}}{Additional arguments passed to read functions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with Metric and Value columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStat-list"></a>}}
\if{latex}{\out{\hypertarget{method-GetStat-list}{}}}
\subsection{Method \code{list()}}{
List tables or files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStat$list(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to read functions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetStat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
