% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer.R
\name{leer}
\alias{leer}
\title{leer - A generic function to read various file formats and database tables}
\usage{
leer(x, ...)
}
\arguments{
\item{x}{the path to the file to read, can be a local file path}

\item{...}{additional arguments passed to the specific read method (e.g., read_csv, read_xpt, etc.) # nolint}
}
\value{
a data frame or tibble containing the contents of the file
}
\description{
leer - A generic function to read various file formats and database tables
}
\examples{
leer("data.csv")
}
