% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dir.R
\name{list_dir}
\alias{list_dir}
\title{List all files in a directory with a specific extension and return
them as a named list.}
\usage{
list_dir(location, ext = "xpt")
}
\arguments{
\item{location}{The directory to search for files.}

\item{ext}{The file extension to filter by (default is "xpt").}
}
\value{
A named list of file paths, where the names are derived from
the file names without extensions.
}
\description{
List all files in a directory with a specific extension and return
them as a named list.
}
\examples{
list_dir("path/to/data", ext = "xpt")
}
