% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwd.R
\name{passwd}
\alias{passwd}
\title{Ask for user password input}
\usage{
passwd(q = "🔑 Password: ")
}
\arguments{
\item{q}{A character string with the prompt message for the password input}
}
\value{
A character string containing the password entered by the user
}
\description{
Ask for user password input
}
\examples{
passwd()
passwd(q="PASSWORD: ")
}
